+++ client/src/components/PlanDetails.js
@@ -44,11 +44,12 @@ function PlanDetails() {
         </Table>
         {currentPlan.length !== 0 ?
             <ListGroup horizontal className="legend">
-                <ListGroup.Item className="flex-fill" variant="secondary">Legend:</ListGroup.Item>
-                <ListGroup.Item className="flex-fill">* Available as UEs</ListGroup.Item>
-                <ListGroup.Item className="flex-fill">~ Available as BDEs</ListGroup.Item>
-                <ListGroup.Item className="flex-fill">^ Self-paced Course</ListGroup.Item>
-                <ListGroup.Item className="flex-fill"># Available as GERPEs</ListGroup.Item>
+            
+                <ListGroup.Item className={`flex-fill ${customOptions.displaySetting}`} variant="secondary">Legend:</ListGroup.Item>
+                <ListGroup.Item className={`flex-fill ${customOptions.displaySetting}`}>* Available as UEs</ListGroup.Item>
+                <ListGroup.Item className={`flex-fill ${customOptions.displaySetting}`}>~ Available as BDEs</ListGroup.Item>
+                <ListGroup.Item className={`flex-fill ${customOptions.displaySetting}`}>^ Self-paced Course</ListGroup.Item>
+                <ListGroup.Item className={`flex-fill ${customOptions.displaySetting}`}># Available as GERPEs</ListGroup.Item>
             </ListGroup> : null
         }
     </>