+++ client/src/services/useWindowDimension.js
@@ -1,6 +1,6 @@
 // credit https://stackoverflow.com/questions/19014250/rerender-view-on-browser-resize-with-react
 
-import React, { useLayoutEffect, useState } from 'react';
+import { useLayoutEffect, useState } from 'react';
 
 export function useWindowDimension() {
   const [size, setSize] = useState([0, 0]);