+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Any new suggestions are welcomed.
 
 ### Added
 
-- legend for symbols in course names
+- legend for symbols in course names in CourseOverview and IndexEditOverview
 
 ### Fixed